% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_email_auth.R
\name{fire_email_auth}
\alias{fire_email_auth}
\title{Authentificate via email with Firebase}
\usage{
fire_email_auth(auth_email, auth_password, api_key)
}
\arguments{
\item{auth_email}{email of the Firebase user}

\item{auth_password}{password  of the Firebase user}

\item{api_key}{web api key of the Firebase database}
}
\value{
list with status and body
}
\description{
Authentificate via email with Firebase
}
