% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_get_token.R
\name{fire_get_token}
\alias{fire_get_token}
\title{Get token to read from Firebase database}
\usage{
fire_get_token(auth_email, auth_password, api_key)
}
\arguments{
\item{auth_email}{email of the Firebase user}

\item{auth_password}{password  of the Firebase user}

\item{api_key}{web api key of the Firebase database}
}
\description{
Get token to read from Firebase database
}
