% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_read_data.R
\name{fire_read_data}
\alias{fire_read_data}
\title{Read data from a Firebase database node in json format}
\usage{
fire_read_data(auth_email, auth_password, api_key, database_url, node)
}
\arguments{
\item{auth_email}{email of the Firebase user}

\item{auth_password}{password  of the Firebase user}

\item{api_key}{web api key of the Firebase database}

\item{database_url}{url of a Firebase database}

\item{node}{path to the Firebase database node}
}
\value{
parsed json into a dataframe
}
\description{
Read data from a Firebase database node in json format
}
